use the same HDAddressPayload
